% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_access_mdb.R
\name{gd_access_mdb}
\alias{gd_access_mdb}
\title{Reads MS Access .mdb files}
\usage{
gd_access_mdb(db_file, read_object = NA)
}
\arguments{
\item{db_file}{The address of the .mdb file, as \code{character}.}

\item{read_object}{Optional. Name of the table or view object to be read, 
as \code{character}. Defaults to \code{NA}. See Value.}
}
\value{
Nothing is returned if \code{read_object = NA} (default), but info on 
non-system objects is printed. When \code{read_object} is specified, a dataframe is returned.
}
\description{
Reads MS Access .mdb files, shows limited info on (non-system) objects, and 
allows reading any such object of type TABLE or VIEW.
}
\examples{
\dontrun{
# list (non-system) objects in MS Access file "my_data.mdb"
gd_access_mdb("my_data.mdb")

# read object "an_object" into dataframe dat
dat <- gd_access_mdb("my_data.mdb", read_object = "an_object")
}


}
