% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_categ_by.R
\name{gd_categ_by}
\alias{gd_categ_by}
\title{Describes a dataframe categorigal variable, by groups}
\usage{
gd_categ_by(df, var, by, ...)
}
\arguments{
\item{df}{The \code{dataframe} containing varibles.}

\item{var}{The categorical variable to be described. Should be either
\code{character} or \code{factor}.}

\item{by}{The categorical variable for grouping. Should be either
\code{character} or \code{factor}.}

\item{...}{Additional arguments to \code{gd_categ()}}
}
\value{
A \code{dataframe} with columns: \code{Variable}, \code{Key}, plus
one additional column for each group in the \code{by} variable.
}
\description{
Describes a dataframe categorigal variable, by groups defined by another
categorical variable in the same dataframe.
}
\examples{
# Example data 
set.seed(123)
dat <- data.frame(Sex = sample(c("Male", "Female"), 100, replace=TRUE), 
                  Age = floor(sample(50 + 10 * rnorm(100))), 
                  Group = sample(LETTERS[1:2], 100, replace=TRUE))
                  
Describe Sex by Group
gd_categ_by(dat, "Sex", "Group")

Same using pipes
library(dplyr)
dat \%>\%
    gd_categ_by("Sex", "Group") 

}
