% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_date.R
\name{gd_date}
\alias{gd_date}
\title{Describes a date variable}
\usage{
gd_date(x, date_format = "\%d-\%b-\%y")
}
\arguments{
\item{x}{The date variable to be described (\code{c("Date", "POSIXlt",
"POSIXct", "POSIXt")}).}

\item{date_format}{Optional. If specified, should ba a valid date format. See
the Details section in \code{\link{strptime}} for more details.}
}
\value{
A \code{dataframe} with three columns: \code{Variable}, \code{Key},
\code{Value}.
}
\description{
Describes a date variable by computing its range.
}
\examples{
# Create example data
inclusion_date <- as.Date(c("2017-01-01", "2017-03-25", "2018-01-31"))

# Default date_format
gd_date(inclusion_date)

# Month as number, and day with leading space for single digit numbers
gd_date(inclusion_date, date_format="\%e-\%m-\%y")

# Four-digit year
gd_date(inclusion_date, date_format="\%e-\%m-\%Y")


}
