% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_date_by.R
\name{gd_date_by}
\alias{gd_date_by}
\title{Describes a dataframe date variable, by groups}
\usage{
gd_date_by(df, var, by, ...)
}
\arguments{
\item{df}{The \code{dataframe} containing varibles.}

\item{var}{The \code{date} variable to be described.}

\item{by}{The categorical variable for grouping. Should be either}

\item{...}{Additional arguments to \code{gd_date()}
\code{character} or \code{factor}.}
}
\value{
A \code{dataframe} with columns: \code{Variable}, \code{Key}, plus
one additional column for each group in the \code{by} variable.
}
\description{
Describes a dataframe date variable, by groups defined by a
categorical variable in the same dataframe.
}
\examples{
# Create example data
inclus_dt <- as.Date(c("2017-01-01", "2017-03-25", "2018-01-31",
                       "2018-01-01", "2018-02-20", "2017-12-07"))
treatment <- rep(LETTERS[1:2],3)
inclusion <- data.frame(inclus_dt, treatment)

gd_date_by(inclusion, "inclus_dt", "treatment")

gd_date_by(inclusion, "inclus_dt", "treatment", "\%e-\%b-\%y")


}
