% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_df_by.R
\name{gd_df_by}
\alias{gd_df_by}
\title{Describes a dataframe by groups}
\usage{
gd_df_by(df, by, ...)
}
\arguments{
\item{df}{\code{dataframe} to be described.}

\item{by}{Name of the variable defining groups. Should be \code{character} or
\code{factor}.}

\item{...}{Further arguments passed to function \code{gd_df}.}
}
\value{
A \code{dataframe} with columns \code{Variable}, \code{Key}, and one
additional column for each level/value of the \code{by} variable.
}
\description{
Describes a dataframe by groups, defined by the \code{by} variable.
}
\examples{
# Example data
set.seed(123)
Sex <- sample(c("Male", "Female"), 100, replace=TRUE)
Age <- floor(sample(50 + 10 * rnorm(100)))
Group <- sample(LETTERS[1:2], 100, replace=TRUE)
dat <- data.frame(Sex, Age, Group)

# Describe dataframe dat by Group
gd_df_by(dat, "Group")

# Same, using pipe
library(dplyr)
dat \%>\% 
  gd_df_by("Group")

# Introduce some missings and pass further arguments
dat$Sex[1:5] <- NA
gd_df_by(dat, "Group")

# Change the descriptor for missings
gd_df_by(dat, "Group", NA_label = "Unknown")
}
