% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_get_decimals.R
\name{gd_get_decimals}
\alias{gd_get_decimals}
\title{Get number of decimals}
\usage{
gd_get_decimals(x)
}
\arguments{
\item{x}{(\code{numeric}) vector.}
}
\value{
A numeric vector, which values are the number of decimals of corresponding \code{x} elements.
}
\description{
Gets the number of decimals for each element of a nuemric vector.
}
