% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_mask.R
\name{gd_mask}
\alias{gd_mask}
\title{Masks numbers in columns of a dataframe}
\usage{
gd_mask(df, col = "Value")
}
\arguments{
\item{df}{\code{dataframe} with column(s) to be masked.}

\item{col}{Columns to be masked. Can be \code{character} or 
\code{numeric}, indicating column names or positions. See examples.}
}
\value{
A \code{dataframe} with same columns as \code{df}, with masked values 
for  \code{cols}.
}
\description{
Masks columns of a dataframe, changing numbers by character "x". This is 
useful to prepare templates for tables in Statistcal Analysis Plans.
}
\examples{
# Example data
set.seed(123) 
Sex <- sample(c("Male", "Female"), 100, replace=TRUE)
Age <- floor(sample(50 + 10 * rnorm(100)))
dat <- data.frame(Sex, Age)

library(dplyr)
dat \%>\%
    gd_df() \%>\%
    gd_mask()

}
