% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_numeric.R
\name{gd_numeric}
\alias{gd_numeric}
\title{Describes a quantitative variable}
\usage{
gd_numeric(x, decimals = NA)
}
\arguments{
\item{x}{The quantitative variable to be described (\code{numeric}).}

\item{decimals}{Optional. Integer controling the rounding (decimals) of
computed statistics. See Details.}
}
\value{
A \code{dataframe} with three columns: \code{Variable}, \code{Key},
\code{Value}.
}
\description{
Describes a quantitative variable by computing common statistics: mean (SD),
median (IRQ), range, n (\%).
}
\details{
If \code{decimals} is not specified, statistics are computed
according to the (median) precision of (\code{x}). Suppose (\code{x}) is
expressed with two decimals. Then, the mean, SD, median and IQR are rounded
to three decimals, while min and max are expressed with two decimals. If
\code{decimals} is specified, mean, SD, median and IQR are rounded to
\code{decimals+1} decimals, while min and max are expressed with \code{decimals}
decimals. However, note that non-significant digits are never printed, e.g.,
17.0 is printed as 17).
}
\examples{
# Example data (first lines)
head(iris)

gd_numeric(iris$Sepal.Length)

# Avoid variable name prefixed with dataframe name
with(iris, gd_numeric(Sepal.Length))

# See Sepal.Length values and note their precision (1 decimal)
iris$Sepal.Length

# Increase precision of statistics
gd_numeric(iris$Sepal.Length, decimals = 2)

# Decrease precision of statistics
Note that min and max are rounded to zero!
gd_numeric(iris$Sepal.Length, decimals = 0)


}
