% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_numeric_by.R
\name{gd_numeric_by}
\alias{gd_numeric_by}
\title{Describes a dataframe numeric variable, by groups}
\usage{
gd_numeric_by(df, var, by, ...)
}
\arguments{
\item{df}{The \code{dataframe} containing varibles.}

\item{var}{The \code{numeric} variable to be described.}

\item{by}{The categorical variable for grouping. Should be either
\code{character} or \code{factor}.}

\item{...}{Additional arguments to \code{gd_numeric()}}
}
\value{
A \code{dataframe} with columns: \code{Variable}, \code{Key}, plus
one additional column for each group in the \code{by} variable.
}
\description{
Describes a dataframe numeric variable, by groups defined by a
categorical variable in the same dataframe.
}
\examples{
gd_numeric_by(iris, "Sepal.Length", "Species")

gd_numeric_by(iris, "Sepal.Length", "Species", decimals=2)


}
