% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_read_MSAccessDB.R
\name{gd_read_MSAccessDB}
\alias{gd_read_MSAccessDB}
\title{Reads MS Access .mdb file}
\usage{
gd_read_MSAccessDB(db_file, types = c("TABLE", "VIEW"))
}
\arguments{
\item{db_file}{\code{character}. The .mbd file address.}

\item{types}{\code{character}. The type of elements to read. More than one 
type can be indicated, as a character vector. Defaults to \code{c("TABLE", 
"VIEW")}.}
}
\value{
The function returs NULL, it is interesting for its side effects: 
creates a dafaframe, in the Global environment, for each database element.
}
\description{
Reads MS Access .mdb file and creates a dataframe for each element (tables 
and views) in the database.
}
\examples{
\dontrun{
db_file <- "C:/Example data/Data.mdb"

# Read all tables and views
gd_read_MSAccessDB(db_file)

# Only tables
gd_read_MSAccessDB(db_file, "TABLE") 

# Only views
gd_read_MSAccessDB(db_file, "VIEW")
}
}
