% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_subset_section.R
\name{gd_subset_section}
\alias{gd_subset_section}
\title{Subsets A Dataframe Selecting Variables Of A CRF section.}
\usage{
gd_subset_section(df, section_name)
}
\arguments{
\item{df}{The dataframe containing all study variables.}

\item{section_name}{Character. Name of the CRF section.}
}
\value{
A \code{dataframe} with all variables in df belonging to section_name.
}
\description{
Subsets the dataframe of all study variables, keepingonly those in specified CRF section. For
this function to work, the dataframe var must exists, and contain variable names in column
VAR, and var labels in column VARIABLE.
}
\examples{
# See sections defined in column SECCION of the var dataframe.
unique(var$SECCION)

# Create dataframe demo, containing only variables from the demographyc data section.
demo <- gd_subset_section(dat, "Cuestionario ACT")
}
