% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_word_table.R
\name{gd_word_table}
\alias{gd_word_table}
\title{Creates a RTF/Word version of standard descriptive result dataframes.}
\usage{
gd_word_table(df, title = "")
}
\arguments{
\item{df}{A dataframe with  columns (\code{Variable}, \code{Key},and \code{Value}) containing
descriptive results.}

\item{title}{Optional. Character to be sed as table title.}
}
\value{
A list (of class `complextable`).
}
\description{
Creates a RTF/Word version of standard dataframes with descriptive results, 
by means of the `flextable` function/package.
}
\examples{
# Example data 
set.seed(123)
Sex <- sample(c("Male", "Female"), 100, replace=TRUE)
Age <- floor(sample(50 + 10 * rnorm(100)))
Group <- sample(LETTERS[1:2], 100, replace=TRUE)
dat <- data.frame(Sex, Age, Group)

# Describe dat and save as result
gd_df(dat) -> result

# Create a word table object 
gd_word_table(result, "Demographic data")

# Same using pipes
library(dplyr)
dat \%>\%
    gd_df() \%>\%
    gd_word_table("Demographic data")

# An example with by-group description
dat \%>\%
    gd_df_by("Group") \%>\% 
    gd_word_table("Demo data")
    
    
}
