% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_write_factor_defs.R
\name{gd_write_factor_defs}
\alias{gd_write_factor_defs}
\title{Writes a script to define factors for coded variables in a dataframe}
\usage{
gd_write_factor_defs(out_file, dat_df = dat, var_df = var,
  var_name = "var_name", var_dictionary = "dictionary", dic_df = dic,
  dic_dictionary = "dictionary", dic_value = "value",
  dic_label = "value_label")
}
\arguments{
\item{out_file}{File name for the script that will be written, as 
\code{character}.}

\item{dat_df}{The \code{dataframe} with coded variables that need to be redefined as 
factors, according to a coding dictionary. Defaults to \code{dat}.}

\item{var_df}{A \code{dataframe} containing info on variables 
in \code{dat_df} and corresponding coding dictionaries. Defaults to 
\code{var}.}

\item{var_name}{The \code{character} name of the column in \code{var_df} 
containing the name of variables in \code{dat_df}. Defaults to 
\code{"var_name"}.}

\item{var_dictionary}{The \code{character} name of the column in 
\code{var_df} containing the name of the coding dictionary of coded 
variables. Defaults to \code{"dictionary"}.}

\item{dic_df}{The \code{dataframe} containing coding dictionaries. Defaults 
to \code{dic}.}

\item{dic_dictionary}{The \code{character} name of the column in 
\code{dic_df} containing the name of dictionaries. Defaults to 
\code{"dictionary"}.}

\item{dic_value}{The \code{character} name of the column in \code{dic_df} 
containing the codes. Defaults to \code{"value"}.}

\item{dic_value}{The \code{character} name of the column in \code{dic_df} 
containing the decodes or descriptors. Defaults to \code{"value_label"}.}
}
\value{
The function returns NULL, but has side effects: it writes file 
\code{out_file} containing the R script necessary to define factors for 
coded variables in \code{dat_df}.
}
\description{
Writes a script to define factors for coded variables in a dataframe, from 
relevant info on the coding dictionaries, contained in auxiliar dataframes. 
The arguments of this function have defaults that assuming a particular names
of relevant objects.
}
\examples{
\dontrun{
# writes definitions into file "factor_definitions.R"
gd_write_factor_defs("factor_definitions.R")
}


}
