% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_df.R
\name{gd_df}
\alias{gd_df}
\title{Describes all variables in a dataframe}
\usage{
gd_df(df, useNA = "ifany", decimals = NA, date_format = "\%d-\%b-\%y")
}
\arguments{
\item{df}{The \code{dataframe} to be described.}

\item{decimals}{Optional. Integer controling the rounding (decimals) of
statistics computed for quantitative variables, as in \code{gd_numeric}.
See \code{?gd_numeric} for details.}

\item{percent}{Optional. If specified, should be one of the following
\code{chacarter} values: \code{"total"} or \code{"valid"}. See details.}
}
\value{
A \code{dataframe} with three columns: \code{Variable}, \code{Key},
\code{Value}.
}
\description{
Describes all variables in a dataframe.
}
\details{
When \code{percent = "total"} (default), missing values are included
and percentages are computed on total number of cases. When
\code{percent = "valid"}, missing values are excluded and percentages are computed
on the number of non-missing cases.
}
\examples{
# Example data (first lines)
head(iris)

# Just to have some missings
iris[sample(1:length(iris)), "Species"] <- NA
iris[sample(1:length(iris)), "Sepal.Length"] <- NA

gd_df(iris)
gd_df(iris, percent = "total")

# Frequency table, excluding missings (valid \%)
gd_df(iris, percent = "valid")
}
