% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_numeric.R
\name{gd_numeric}
\alias{gd_numeric}
\title{Describes a numeric variable}
\usage{
gd_numeric(x, decimals = NA, language = "sp")
}
\arguments{
\item{x}{\code{numeric}. Variable to be described.}

\item{decimals}{Optional. Controls the rounding of summary statistics. If 
specified, should be \code{numeric (integer)}. See Details.}

\item{language}{\code{character}. Language for descriptive statistic 
descriptors. Should be on of "sp", "en" or "ca". Defaults to "sp".}
}
\value{
A \code{dataframe} with three columns: \code{Variable}, \code{Key},
\code{Value}.
}
\description{
Describes a numeric variable by computing common summary statistics.
}
\details{
Statistics are computed with adequate precision (rounding) in each 
case: mean, median, sandard deviation and IQR values are rounded to one 
additional decimal wrt to \code{x} values, while extremes (min, max) are 
rounded to the same precision as \code{x} values. This will be convenient in 
most cases, but may be changed by specifying \code{decimals}. If specified, 
the mean, median, SD and IQR values are rounded to \code{decimals+1}, while 
extremes are rounded to \code{decimals}. See examples.  

In any case, note that non-significant digits are not printed (e.g.,17.0 is 
printed as 17).
}
\examples{
# Example data 
weight <- seq(60,95,5)
height <- 100 + weight
bmi <- weight/(height/100)^2

gd_numeric(weight)
gd_numeric(height, language = "ca")

# Excessive precission due to lack of rounding in computed bmi
gd_numeric(bmi)

# Reduce precision
gd_numeric(bmi, decimals = 1)

}
